… docs/compressing_a_raw_clip.md
… docs/error_metrics.md
… includes/acl/compression/utils.h
… includes/acl/core/impl/compiler_utils.h
… includes/acl/core/track_types.h
… includes/acl/decompression/impl/decompress_data.h
… includes/acl/decompression/impl/track_sampling_impl.h
… includes/acl/math/scalar_packing.h
… includes/acl/math/vector4_packing.h
… test_data/configs/uniformly_sampled_quant_bind_relative.config.sjson
… test_data/configs/uniformly_sampled_quant_mtx_error.config.sjson
… tests/sources/core/test_bitset.cpp
… tests/sources/math/test_scalar_packing.cpp
… tests/sources/math/test_vector4_packing.cpp
… tools/acl_compressor/acl_compressor.py
… tools/acl_compressor/sources/acl_compressor.cpp
… tools/acl_decompressor/acl_decompressor.py
… tools/calc_local_bit_rates.py
… tools/graph_generation/gen_bit_rate_stats.py
… tools/graph_generation/gen_decomp_delta_stats.py
… tools/graph_generation/gen_decomp_stats.py
… tools/graph_generation/gen_full_error_stats.py
… tools/graph_generation/gen_summary_stats.py
… tools/graph_generation/pack_png.py
… tools/release_scripts/extract_stats.py
… tools/release_scripts/test_everything.py