I measured before and after this PR on my Ryzen 2950X desktop. Here is the average compression speed of 3 runs for CMU, Paragon, and Fortnite as well as the 50th, 85th, 99th percentile for the compression time of their clips. Compression is ~2x faster across the board.

Average compression speed  | Before | After
--- | --- | ---
CMU | 10682.45 KB/sec | 19788.26 KB/sec
Paragon | 7439.9 KB/sec | 14636.82 KB/sec
Fortnite | 14845.71 KB/sec | 28561.8 KB/sec

CMU compression time  | Before | After
--- | --- | ---
50th percentile | 0.029 sec | 0.015 sec
85th percentile | 0.113 sec | 0.059 sec
99th percentile | 0.300 sec | 0.158 sec

Paragon compression time  | Before | After
--- | --- | ---
50th percentile | 0.029 sec | 0.015 sec
85th percentile | 0.123 sec | 0.062 sec
99th percentile | 1.368 sec | 0.673 sec

Fortnite compression time  | Before | After
--- | --- | ---
50th percentile | 0.015 sec | 0.008 sec
85th percentile | 0.229 sec | 0.120 sec
99th percentile | 1.943 sec | 0.981 sec